+++ arm9/source/graphics/FontGraphic.h
@@ -33,7 +33,7 @@ public:
 	FontGraphic() { };
 	int load(int textureID, glImage *_font_sprite,
 			const unsigned int numframes,
-			const unsigned int *texcoords,
+			const u16 *texcoords,
 			GL_TEXTURE_TYPE_ENUM type,
 			int sizeX,
 			int sizeY,
